frontend: Switch from react to Svelte
I've been keen to try Svelte for a while based on the likely large
reduction in total bundle size and performance touted. The lack of
typesript support has put me off - it seems ok now, but note that
linting doesn't work. Note also that this site is very static, and is
built this way to allow me to experiment with features.

Sapper with export is used and I want to use Quart as the server,
rather than anything built in js.