frontend: Upgrade the dependences
highlightjs-svelte has stopped working as an importable package, so
I've vendored the code and imported it within the project.

With the switch to vite, which has a proxy feature, the unreliable
hook can be removed (although the file is required). However it is not
possible to pre-render with the static images being proxied.